requirements.txt
telepot==12.7
yt-dlp==2024.3.10
python-dotenv==1.0.1
loguru==0.7.2
instaloader==4.14.1 